MemeComp/requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
click==8.1.3
colorama==0.4.6
constantly==15.1.0
cryptography==40.0.2
cssbeautifier==1.14.7
daphne==4.0.0
Django==4.2.1
django-nine==0.2.7
djangorestframework==3.14.0
djlint==1.27.0
EditorConfig==0.12.3
html-tag-names==0.1.2
html-void-elements==0.1.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
jsbeautifier==1.14.7
json5==0.9.11
msgpack==1.0.5
packaging==23.1
pathspec==0.11.1
Pillow==9.5.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==23.1.1
pytz==2023.3
PyYAML==6.0
redis==4.5.4
regex==2023.5.5
service-identity==21.1.0
six==1.16.0
ska==1.9.1
sqlparse==0.4.4
tomli==2.0.1
tqdm==4.65.0
Twisted==22.10.0
twisted-iocpsupport==1.0.3
txaio==23.1.1
typing_extensions==4.5.0
tzdata==2023.3
zope.interface==6.0
